Elastic Changes and Tests

+++ api/v1/lib/Oxzion/test/AnalyticsTest/AnalyticsTest.php
@@ -27,9 +27,12 @@ class AnalyticsTest extends MainControllerTest{
         $this->loadConfig();
         parent::setUp();
         $this->setSearchData();
+        
         $config = $this->getApplicationConfig();
         $this->searchFactory = new SearchFactory($config);   
-        $this->analyticsFactory = new AnalyticsFactory($config);   
+        $this->analyticsFactory = new AnalyticsFactory($config);  
+        $this->setupData(); 
+        sleep (1) ;
     }   
     
 
@@ -38,29 +41,25 @@ class AnalyticsTest extends MainControllerTest{
           $this->dataset = $parser->parseYaml(dirname(__FILE__)."/Dataset/Analytics.yml");
     }
 
-    public function assertIndex($indexer,$body) {       
+    public function createIndex($indexer,$body) {       
         $type = 'type';
         $app_id = $body['app_id'];
         $id = $body['id'];
         AuthContext::put(AuthConstants::ORG_ID, $body['org_id']);
         $return=$indexer->index($app_id,$id,$type,$body);
-        $this->assertEquals($return['result'],"created");
     }
 
-    public function testIndex(){
-        if(enableElastic==0){
-            $this->markTestSkipped('Only Integration Test');        
+    public function setupData(){
+        if(enableElastic!=0){ 
+            $indexer = $this->searchFactory->getIndexer();
+            $dataset = $this->dataset['ox_analysis'];
+            foreach($dataset as $body) {
+                $this->createIndex($indexer,$body);
+            }
         }
-        $indexer = $this->searchFactory->getIndexer();
-        $body = $this->dataset['ox_search'];
-        $this->assertIndex($indexer,$body[0]);
-        $this->assertIndex($indexer,$body[1]);
-        $this->assertIndex($indexer,$body[2]);
-        $this->assertIndex($indexer,$body[3]);
-        $this->assertIndex($indexer,$body[4]);
-        $this->assertIndex($indexer,$body[5]);
     }
 
+    
     public function testGrouping() {
         if(enableElastic==0){
             $this->markTestSkipped('Only Integration Test');        
@@ -69,6 +68,7 @@ class AnalyticsTest extends MainControllerTest{
         $ae = $this->analyticsFactory->getAnalyticsEngine();
         $parameters = ['group'=>'created_by','field'=>'amount','operation'=>'sum','date-period'=>'2018-01-01/2018-12-12','date_type'=>'date_created'];
         $results = $ae->runQuery('11_test',null,$parameters);
+        $results = $results['data'];
         $this->assertEquals($results[0]['name'], "John Doe");
         $this->assertEquals($results[0]['value'], "800");
         $this->assertEquals($results[0]['grouplist'], "created_by");
@@ -85,6 +85,7 @@ class AnalyticsTest extends MainControllerTest{
         $ae = $this->analyticsFactory->getAnalyticsEngine();
         $parameters = ['group'=>'created_by,category','field'=>'amount','operation'=>'sum','date-period'=>'2018-01-01/2018-12-12','date_type'=>'date_created'];
         $results = $ae->runQuery('11_test',null,$parameters);
+        $results = $results['data'];
         $this->assertEquals($results[0]['name'], "John Doe - A");
         $this->assertEquals($results[0]['value'], "200");
         $this->assertEquals($results[0]['grouplist'][0], "John Doe");
@@ -93,16 +94,29 @@ class AnalyticsTest extends MainControllerTest{
         $this->assertEquals($results[2]['grouplist'][0], "Mike Price");
     }
 
-
-    public function testDelete() {
+    public function testLists() {
         if(enableElastic==0){
             $this->markTestSkipped('Only Integration Test');        
         }
-        $indexer = $this->searchFactory->getIndexer();
-        $return1=$indexer->delete('11_test','all');
-        $return2=$indexer->delete('12_test','all');
-        $this->assertEquals($return1['acknowledged'],1);
-        $this->assertEquals($return2['acknowledged'],1);
+        AuthContext::put(AuthConstants::ORG_ID, 1);
+        $ae = $this->analyticsFactory->getAnalyticsEngine();
+        $parameters = ['field'=>'amount','date-period'=>'2018-01-01/2019-12-12','date_type'=>'date_created','list'=>'name,created_by,category'];
+        $results = $ae->runQuery('11_test',null,$parameters);
+        $results = $results['data'];
+        $this->assertEquals($results[0]['name'], "testing document");
+        $this->assertEquals($results[0]['category'], "A");
+        $this->assertEquals($results[0]['created_by'], "Mike Price");                
+    }
+
+
+    public function tearDown()
+    {
+         parent::tearDown();
+         if(enableElastic!=0){                
+            $indexer = $this->searchFactory->getIndexer();
+            $return1=$indexer->delete('11_test','all');
+            $return2=$indexer->delete('12_test','all');
+        }    
     }
